+++ Controller/Index/Push.php
@@ -1,0 +1,95 @@
+<?php
+
+namespace Heidelpay\Gateway\Controller\Index;
+
+use Heidelpay\Gateway\Helper\Payment as HeidelpayHelper;
+use Magento\Sales\Model\Order\Email\Sender\InvoiceSender;
+use Magento\Sales\Model\Order\Email\Sender\OrderCommentSender;
+use Magento\Sales\Model\Order\Email\Sender\OrderSender;
+
+/**
+ * Summary
+ *
+ * Desc
+ *
+ * @license Use of this software requires acceptance of the License Agreement. See LICENSE file.
+ * @copyright Copyright © 2016-present Heidelberger Payment GmbH. All rights reserved.
+ * @link https://dev.heidelpay.de/magento2
+ *
+ * @author Stephano Vogel
+ *
+ * @package heidelpay
+ * @subpackage magento2
+ * @category magento2
+ */
+class Push extends \Heidelpay\Gateway\Controller\HgwAbstract
+{
+    /**
+     * @var \Heidelpay\PhpApi\Push
+     */
+    protected $heidelpayPush;
+
+    /**
+     * Push constructor.
+     *
+     * @param \Magento\Framework\App\Action\Context $context
+     * @param \Magento\Customer\Model\Session $customerSession
+     * @param \Magento\Checkout\Model\Session $checkoutSession
+     * @param \Magento\Sales\Model\OrderFactory $orderFactory
+     * @param \Magento\Framework\Url\Helper\Data $urlHelper
+     * @param \Psr\Log\LoggerInterface $logger
+     * @param \Magento\Quote\Api\CartManagementInterface $cartManagement
+     * @param \Magento\Quote\Api\CartRepositoryInterface $quoteObject
+     * @param \Magento\Framework\View\Result\PageFactory $resultPageFactory
+     * @param HeidelpayHelper $paymentHelper
+     * @param OrderSender $orderSender
+     * @param InvoiceSender $invoiceSender
+     * @param OrderCommentSender $orderCommentSender
+     * @param \Magento\Framework\Encryption\Encryptor $encryptor
+     * @param \Magento\Customer\Model\Url $customerUrl
+     * @param \Heidelpay\PhpApi\Push $heidelpayPush
+     */
+    public function __construct(
+        \Magento\Framework\App\Action\Context $context,
+        \Magento\Customer\Model\Session $customerSession,
+        \Magento\Checkout\Model\Session $checkoutSession,
+        \Magento\Sales\Model\OrderFactory $orderFactory,
+        \Magento\Framework\Url\Helper\Data $urlHelper,
+        \Psr\Log\LoggerInterface $logger,
+        \Magento\Quote\Api\CartManagementInterface $cartManagement,
+        \Magento\Quote\Api\CartRepositoryInterface $quoteObject,
+        \Magento\Framework\View\Result\PageFactory $resultPageFactory,
+        HeidelpayHelper $paymentHelper,
+        OrderSender $orderSender,
+        InvoiceSender $invoiceSender,
+        OrderCommentSender $orderCommentSender,
+        \Magento\Framework\Encryption\Encryptor $encryptor,
+        \Magento\Customer\Model\Url $customerUrl,
+        \Heidelpay\PhpApi\Push $heidelpayPush
+    ) {
+        parent::__construct(
+            $context,
+            $customerSession,
+            $checkoutSession,
+            $orderFactory,
+            $urlHelper,
+            $logger,
+            $cartManagement,
+            $quoteObject,
+            $resultPageFactory,
+            $paymentHelper,
+            $orderSender,
+            $invoiceSender,
+            $orderCommentSender,
+            $encryptor,
+            $customerUrl
+        );
+
+        $this->heidelpayPush = $heidelpayPush;
+    }
+
+    public function execute()
+    {
+        $this->_logger->debug('Params from Response: ' . json_encode($this->getRequest()->getParams()));
+    }
+}